+++ src/especialidades/especialidades.service.ts
@@ -2,6 +2,7 @@ import { orm } from "../../shared/orm.js";
 import { Service } from "../../shared/service.js";
 import { ObjectId } from "mongodb";
 import { Especialidades } from "./especialidades.entity.js";
+import { NotFound } from "../../shared/errors.js";
 
 const em = orm.em;
 
@@ -13,13 +14,17 @@ export class EspecialidadesService implements Service<Especialidades> {
   public async findOne(item: {
     id: string;
   }): Promise<Especialidades | undefined> {
-    return await em.findOneOrFail(
+    const especialidad = await em.findOne(
       Especialidades,
       {
         _id: new ObjectId(item.id),
       },
       { populate: ["medicos"] }
     );
+
+    if (!especialidad) throw new NotFound(item.id);
+
+    return especialidad;
   }
 
   public async add(item: Especialidades): Promise<Especialidades | undefined> {
@@ -31,22 +36,30 @@ export class EspecialidadesService implements Service<Especialidades> {
   public async update(
     item: Especialidades
   ): Promise<Especialidades | undefined> {
-    const especialidadAActualizar = await em.findOneOrFail(Especialidades, {
+    const especialidadAActualizar = await em.findOne(Especialidades, {
       _id: new ObjectId(item.id),
     });
+
+    if (!especialidadAActualizar) throw new NotFound(item.id);
+
     const especialidadActualizada = em.assign(especialidadAActualizar, item);
     await em.flush();
+
     return especialidadActualizada;
   }
 
   public async remove(item: {
     id: string;
   }): Promise<Especialidades | undefined> {
-    const EspecialidadesABorrar = await em.findOneOrFail(Especialidades, {
+    const EspecialidadesABorrar = await em.findOne(Especialidades, {
       _id: new ObjectId(item.id),
     });
+
+    if (!EspecialidadesABorrar) throw new NotFound(item.id);
+
     em.remove(EspecialidadesABorrar);
     await em.flush();
+
     return EspecialidadesABorrar;
   }
 }
