+++ src/medicos/medicos.service.ts
@@ -1,10 +1,11 @@
 import { orm } from "../../shared/orm.js";
-import { Medicos } from "../medicos/medicos.entity.js";
+import { Medicos } from "./medicos.entity.js";
 import { EntityManager, EntityRepository } from "@mikro-orm/core";
 import { Service } from "../../shared/service.js";
 import { ObjectId } from "mongodb";
 import { PopulateHint } from "@mikro-orm/mongodb";
 import { Especialidades } from "../especialidades/especialidades.entity.js";
+import { NotFound } from "../../shared/errors.js";
 
 const em = orm.em;
 
@@ -14,11 +15,15 @@ export class MedicoService implements Service<Medicos> {
   }
 
   public async findOne(item: { id: string }): Promise<Medicos | undefined> {
-    return await em.findOneOrFail(
+    const medico = await em.findOne(
       Medicos,
       { _id: new ObjectId(item.id) },
       { populate: ["especialidad"] }
     );
+
+    if (!medico) throw new NotFound(item.id);
+
+    return medico;
   }
 
   public async add(item: Medicos): Promise<Medicos | undefined> {
@@ -37,20 +42,26 @@ export class MedicoService implements Service<Medicos> {
   }
 
   public async update(item: Medicos): Promise<Medicos | undefined> {
-    const medicoAActualizar = await em.findOneOrFail(Medicos, {
-      _id: new ObjectId(item.id),
-    });
+    const medicoAActualizar = em.getReference(Medicos, new ObjectId(item.id));
+
+    if (!medicoAActualizar) throw new NotFound(item.id);
+
     const medicoActualizado = em.assign(medicoAActualizar, item);
     await em.flush();
+
     return medicoActualizado;
   }
 
   public async remove(item: { id: string }): Promise<Medicos | undefined> {
-    const medicosABorrar = await em.findOneOrFail(Medicos, {
+    const medicosABorrar = await em.findOne(Medicos, {
       _id: new ObjectId(item.id),
     });
+
+    if (!medicosABorrar) throw new NotFound(item.id);
+
     em.remove(medicosABorrar);
     await em.flush();
+
     return medicosABorrar;
   }
 }
