+++ src/pacientes/pacientes.service.ts
@@ -1,7 +1,8 @@
 import { orm } from "../../shared/orm.js";
-import { Pacientes } from "../pacientes/pacientes.entity.js";
+import { Pacientes } from "./pacientes.entity.js";
 import { Service } from "../../shared/service.js";
 import { ObjectId } from "mongodb";
+import { NotFound } from "../../shared/errors.js";
 
 const em = orm.em;
 
@@ -11,7 +12,13 @@ export class PacienteService implements Service<Pacientes> {
   }
 
   public async findOne(item: { id: string }): Promise<Pacientes | undefined> {
-    return await em.findOneOrFail(Pacientes, { _id: new ObjectId(item.id) });
+    const paciente = await em.findOne(Pacientes, {
+      _id: new ObjectId(item.id),
+    });
+
+    if (!paciente) throw new NotFound(item.id);
+
+    return paciente;
   }
 
   public async add(item: Pacientes): Promise<Pacientes | undefined> {
@@ -21,11 +28,15 @@ export class PacienteService implements Service<Pacientes> {
   }
 
   public async update(item: Pacientes): Promise<Pacientes | undefined> {
-    const pacienteAActualizar = await em.findOneOrFail(Pacientes, {
+    const pacienteAActualizar = await em.findOne(Pacientes, {
       _id: new ObjectId(item.id),
     });
+
+    if (!pacienteAActualizar) throw new NotFound(item.id);
+
     const pacienteActualizado = em.assign(pacienteAActualizar, item);
     await em.flush();
+
     return pacienteActualizado;
   }
 
@@ -33,8 +44,12 @@ export class PacienteService implements Service<Pacientes> {
     const pacienteABorrar = await em.findOneOrFail(Pacientes, {
       _id: new ObjectId(item.id),
     });
+
+    if (!pacienteABorrar) throw new NotFound(item.id);
+
     em.remove(pacienteABorrar);
     await em.flush();
+
     return pacienteABorrar;
   }
 }
